Add support for multiple compose.ValueSet and compose.System.
